Release working version
Right now the program runs. There is an input_test.xlsx file used for
testing the output of the program, which writes the "results" to the
output.txt file.

You can run the code for example:

        python cli.py --input_file=input_test.xlsx --solver=glpk

This runs the program on input_test.xlsx using the glpk solver and
outputs the results into output.txt.