Add new way to create constraints
This now uses a function decorator to mark a function as a constraint.
The function that is marked by the constraint decorator is essentially
the rule function that is being passed to Pyomo's Constraint.

This is so that new constraints can be easily added and in exactly one
place. There is no more need to add both the function and call the
function in another function that adds all the constraints.